requirements.txt
fastapi
uvicorn
pydantic
networkx
matplotlib
python-dotenv
google-genai
google-adk
